restructured and clean up
