+++ src/components/columns/navigator/index.tsx
@@ -1,11 +1,11 @@
 import ActiveLink from '@components/activeLink';
-import CreatePost from '@components/navigatorBar/createPost';
-import { ProfileMenu } from '@components/navigatorBar/profileMenu';
+import CreatePost from '@components/columns/navigator/createPost';
+import { UserDropdownMenu } from '@components/columns/navigator/userDropdownMenu';
 
 import { PlusIcon } from '@radix-ui/react-icons';
 import { useLocalStorage } from '@rehooks/local-storage';
 
-export default function NavigatorBar() {
+export default function NavigatorColumn() {
   const [currentUser]: any = useLocalStorage('current-user');
   const profile = JSON.parse(currentUser.metadata);
 
@@ -17,7 +17,7 @@ export default function NavigatorBar() {
           <div className="flex flex-col p-2">
             <div className="flex items-center justify-between">
               <h5 className="font-semibold leading-tight text-zinc-100">{profile.display_name || ''}</h5>
-              <ProfileMenu pubkey={currentUser.pubkey} />
+              <UserDropdownMenu pubkey={currentUser.pubkey} />
             </div>
             <span className="text-sm leading-tight text-zinc-500">@{profile.username || ''}</span>
           </div>
@@ -38,7 +38,7 @@ export default function NavigatorBar() {
           </div>
           <div className="flex flex-col gap-1 text-zinc-500">
             <ActiveLink
-              href={`/feed/following`}
+              href={`/newsfeed/following`}
               activeClassName="ring-1 ring-white/10 dark:bg-zinc-900 dark:text-white"
               className="flex h-10 items-center gap-1 rounded-lg px-2.5 text-sm font-medium hover:bg-zinc-900"
             >
@@ -46,7 +46,7 @@ export default function NavigatorBar() {
               <span>following</span>
             </ActiveLink>
             <ActiveLink
-              href={`/feed/global`}
+              href={`/newsfeed/global`}
               activeClassName="ring-1 ring-white/10 dark:bg-zinc-900 dark:text-white"
               className="flex h-10 items-center gap-1 rounded-lg px-2.5 text-sm font-medium hover:bg-zinc-900"
             >
