+++ src/pages/newsfeed/circle.tsx
@@ -1,5 +1,5 @@
-import BaseLayout from '@layouts/baseLayout';
-import NewsFeedLayout from '@layouts/newsfeedLayout';
+import BaseLayout from '@layouts/base';
+import NewsFeedLayout from '@layouts/newsfeed';
 
 import { JSXElementConstructor, ReactElement, ReactFragment, ReactPortal } from 'react';
