+++ src/pages/onboarding/login/index.tsx
@@ -1,5 +1,5 @@
-import BaseLayout from '@layouts/baseLayout';
-import OnboardingLayout from '@layouts/onboardingLayout';
+import BaseLayout from '@layouts/base';
+import OnboardingLayout from '@layouts/onboarding';
 
 import { motion } from 'framer-motion';
 import { useRouter } from 'next/router';
